# backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/hotel-packages/{providerId}/packages/{packageId}": {
      "patch": {
        "operationId": "HotelPackagesController_update",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "packageId",
            "required": true,
            "in": "path",
            "description": "HotelPackage id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHotelPackageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hotel package updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a hotel package (DRAFT only)",
        "tags": ["hotel-packages"]
      },
      "delete": {
        "operationId": "HotelPackagesController_remove",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "packageId",
            "required": true,
            "in": "path",
            "description": "HotelPackage id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel package deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a hotel package (DRAFT only)",
        "tags": ["hotel-packages"]
      }
    },
...
    "/v1/tour-packages": {
      "get": {
        "operationId": "TourPackagesController_list",
        "parameters": [
          {
            "name": "providerId",
            "required": false,
            "in": "query",
            "description": "Filter by provider ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": ["DRAFT", "PUBLISHED", "PAUSED", "ARCHIVED"]
            }
          },
          {
            "name": "tripType",
            "required": false,
            "in": "query",
            "description": "Filter by trip type",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by name (ILIKE)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "description": "Include archived packages",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": ["createdAt", "name"]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": ["asc", "desc"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tour packages"
          }
        },
        "summary": "List tour packages",
        "tags": ["tour-packages"]
      }
    },
    "/v1/tour-packages/{id}": {
      "get": {
        "operationId": "TourPackagesController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tour package detail"
          }
        },
        "summary": "Get tour package by id",
        "tags": ["tour-packages"]
      }
    },
    "/v1/tour-packages/{providerId}/packages": {
      "post": {
        "operationId": "TourPackagesController_createTourPackage",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTourPackageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tour package created in DRAFT status"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create tour package",
        "tags": ["tour-packages"]
      }
    },
    "/v1/tour-packages/{providerId}/packages/{packageId}/publish": {
      "post": {
        "operationId": "TourPackagesController_publish",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "packageId",
            "required": true,
            "in": "path",
            "description": "TourPackage id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tour package published"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish a tour package (requires verified provider)",
        "tags": ["tour-packages"]
      }
    },
    "/v1/tour-packages/{providerId}/packages/{packageId}/pause": {
      "post": {
        "operationId": "TourPackagesController_pause",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "packageId",
            "required": true,
            "in": "path",
            "description": "TourPackage id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tour package paused"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pause a published tour package",
        "tags": ["tour-packages"]
      }
    },
    "/v1/tour-packages/{providerId}/packages/{packageId}/archive": {
      "post": {
        "operationId": "TourPackagesController_archive",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "packageId",
            "required": true,
            "in": "path",
            "description": "TourPackage id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tour package archived"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive a tour package",
        "tags": ["tour-packages"]
      }
    },
    "/v1/tour-packages/{providerId}/packages/step1": {
      "post": {
        "operationId": "TourPackagesController_step1TripType",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Step1TripTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tour package created in DRAFT status"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Step 1: Set trip type (creates DRAFT package)",
        "tags": ["tour-packages"]
      }
    },
    "/v1/tour-packages/{providerId}/packages/{id}/step2-basics": {
      "patch": {
        "operationId": "TourPackagesController_step2Basics",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Step2BasicsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Step 2: Set basics",
        "tags": ["tour-packages"]
      }
    },
    "/v1/tour-packages/{providerId}/packages/{id}/step3-departures": {
      "put": {
        "operationId": "TourPackagesController_step3Departures",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Step3DeparturesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Step 3: Set departures (replace)",
        "tags": ["tour-packages"]
      }
    },
    "/v1/tour-packages/{providerId}/packages/{id}/step4-pickups": {
      "put": {
        "operationId": "TourPackagesController_step4Pickups",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Step4PickupsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Step 4: Set pickups (replace)",
        "tags": ["tour-packages"]
      }
    },
    "/v1/tour-packages/{providerId}/packages/{id}/step5-highlights": {
      "patch": {
        "operationId": "TourPackagesController_step5Highlights",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Step5HighlightsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Step 5: Set highlights",
        "tags": ["tour-packages"]
      }
    },
    "/v1/tour-packages/{providerId}/packages/{id}/step6-itinerary": {
      "put": {
        "operationId": "TourPackagesController_step6Itinerary",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Step6ItineraryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Step 6: Set itinerary (replace)",
        "tags": ["tour-packages"]
      }
    },
    "/v1/tour-packages/{providerId}/packages/{id}/step7-inclusions-exclusions": {
      "patch": {
        "operationId": "TourPackagesController_step7InclusionsExclusions",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "ProviderProfile id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Step7InclusionsExclusionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Step 7: Set inclusions and exclusions",
        "tags": ["tour-packages"]
      }
    },
...
  },
  "info": {
    "title": "TripAvail API",
    "description": "Two-sided travel marketplace API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4100",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateHotelPackageDto": {
        "type": "object",
        "properties": {
          "templateId": {
            "type": "string",
            "example": "weekend-getaway",
            "enum": [
              "weekend-getaway",
              "honeymoon",
              "family-vacation",
              "corporate-retreat",
              "ski-snow-holiday",
              "golf-getaway",
              "staycation",
              "wellness-spa",
              "cultural-experience",
              "food-wine",
              "anniversary",
              "eco-nature",
              "festival-events",
              "adventure"
            ]
          },
          "name": {
            "type": "string",
            "example": "Romantic Weekend"
          },
          "description": {
            "type": "string",
            "example": "Two nights with breakfast and spa access."
          },
          "pricePerPerson": {
            "type": "number",
            "example": 399.99
          },
          "inclusions": {
            "example": ["Breakfast", "Spa access"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "exclusions": {
            "example": ["Flights not included"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "availabilityRule": {
            "type": "string",
            "example": "WEEKEND_ONLY",
            "enum": ["WEEKEND_ONLY", "SEASONAL", "FLEXIBLE"]
          },
          "amenityIds": {
            "example": ["amenity_cuid1", "amenity_cuid2"],
            "description": "Replace amenities on the package (DRAFT only)",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateTourPackageDto": {
        "type": "object",
        "properties": {
          "tripType": {
            "type": "string",
            "example": "ADVENTURE"
          },
          "name": {
            "type": "string",
            "example": "Island Hopping"
          },
          "description": {
            "type": "string",
            "example": "3-day guided tour across islands."
          },
          "duration": {
            "type": "number",
            "example": 3
          },
          "basePrice": {
            "type": "number",
            "example": 499.5
          },
          "maxSeats": {
            "type": "number",
            "example": 20
          }
        },
        "required": ["tripType", "name", "description", "duration", "basePrice", "maxSeats"]
      },
      "Step1TripTypeDto": {
        "type": "object",
        "properties": {
          "tripType": {
            "type": "string",
            "example": "Adventure"
          }
        },
        "required": ["tripType"]
      },
      "Step2BasicsDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Rocky Mountain Hiking Expedition"
          },
          "description": {
            "type": "string",
            "example": "5-day guided hiking adventure through the Rockies."
          },
          "duration": {
            "type": "number",
            "example": 5,
            "description": "Number of days"
          },
          "basePrice": {
            "type": "number",
            "example": 1299.99
          },
          "maxSeats": {
            "type": "number",
            "example": 12
          }
        },
        "required": ["name", "description", "duration", "basePrice", "maxSeats"]
      },
      "Step3DeparturesDto": {
        "type": "object",
        "properties": {
          "departureDates": {
            "example": ["2026-02-15", "2026-03-15"],
            "description": "ISO dates (YYYY-MM-DD)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "availableSeats": {
            "type": "number",
            "example": 12
          }
        },
        "required": ["departureDates", "availableSeats"]
      },
      "PickupInputDto": {
        "type": "object",
        "properties": {
          "city": {
            "type": "string",
            "example": "Denver"
          },
          "location": {
            "type": "string",
            "example": "Denver International Airport - Arrivals Gate"
          },
          "latitude": {
            "type": "number",
            "example": 39.8561
          },
          "longitude": {
            "type": "number",
            "example": -104.6737
          }
        },
        "required": ["city", "location", "latitude", "longitude"]
      },
      "Step4PickupsDto": {
        "type": "object",
        "properties": {
          "pickups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PickupInputDto"
            }
          }
        },
        "required": ["pickups"]
      },
      "Step5HighlightsDto": {
        "type": "object",
        "properties": {
          "highlights": {
            "example": ["Summit hike to Eagle Peak", "Alpine Lake swim"],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["highlights"]
      },
      "ItineraryDayInputDto": {
        "type": "object",
        "properties": {
          "day": {
            "type": "number",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Arrival and Base Camp Setup"
          },
          "description": {
            "type": "string",
            "example": "Meet at the trailhead, gear check, and hike to base camp."
          }
        },
        "required": ["day", "title", "description"]
      },
      "Step6ItineraryDto": {
        "type": "object",
        "properties": {
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItineraryDayInputDto"
            }
          }
        },
        "required": ["days"]
      },
      "Step7InclusionsExclusionsDto": {
        "type": "object",
        "properties": {
          "inclusions": {
            "example": ["Guided hikes", "Camping gear"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "exclusions": {
            "example": ["Flights", "Travel insurance"],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["inclusions", "exclusions"]
      }
    }
  }
}